fix overflow error measurement

+++ benchmark/disjoint_set/disjoint_set_benchmark.cc
@@ -15,21 +15,21 @@ std::ofstream DisjointSetBenchmark::rank_union_PS_statistics_;
 std::ofstream DisjointSetBenchmark::rank_union_PH_statistics_;
 
 TEST_F(DisjointSetBenchmark, QuickUnion) {
-  RunBenchmark<disjoint_set::QuickUnion<uint16_t>>(quick_union_statistics_);
+  RunBenchmark<disjoint_set::QuickUnion<uint64_t>>(quick_union_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, QuickUnionFC) {
-  RunBenchmark<disjoint_set::QuickUnionFullCompression<uint16_t>>(
+  RunBenchmark<disjoint_set::QuickUnionFullCompression<uint64_t>>(
       quick_union_FC_statistics_, true);
 }
 
 TEST_F(DisjointSetBenchmark, QuickUnionPS) {
-  RunBenchmark<disjoint_set::QuickUnionPathSplitting<uint16_t>>(
+  RunBenchmark<disjoint_set::QuickUnionPathSplitting<uint64_t>>(
       quick_union_PS_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, QuickUnionPH) {
-  RunBenchmark<disjoint_set::QuickUnionPathHalving<uint16_t>>(
+  RunBenchmark<disjoint_set::QuickUnionPathHalving<uint64_t>>(
       quick_union_PH_statistics_);
 }
 
@@ -38,36 +38,36 @@ TEST_F(DisjointSetBenchmark, WeightUnion) {
 }
 
 TEST_F(DisjointSetBenchmark, WeightUnionFC) {
-  RunBenchmark<disjoint_set::WeightUnionFullCompression<int16_t>>(
+  RunBenchmark<disjoint_set::WeightUnionFullCompression<int64_t>>(
       weight_union_FC_statistics_, true);
 }
 
 TEST_F(DisjointSetBenchmark, WeightUnionPS) {
-  RunBenchmark<disjoint_set::WeightUnionPathSplitting<int16_t>>(
+  RunBenchmark<disjoint_set::WeightUnionPathSplitting<int64_t>>(
       weight_union_PS_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, WeightUnionPH) {
-  RunBenchmark<disjoint_set::WeightUnionPathHalving<int16_t>>(
+  RunBenchmark<disjoint_set::WeightUnionPathHalving<int64_t>>(
       weight_union_PH_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, RankUnion) {
-  RunBenchmark<disjoint_set::RankUnion<int16_t>>(rank_union_statistics_);
+  RunBenchmark<disjoint_set::RankUnion<int64_t>>(rank_union_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, RankUnionFC) {
-  RunBenchmark<disjoint_set::RankUnionFullCompression<int16_t>>(
+  RunBenchmark<disjoint_set::RankUnionFullCompression<int64_t>>(
       rank_union_FC_statistics_, true);
 }
 
 TEST_F(DisjointSetBenchmark, RankUnionPS) {
-  RunBenchmark<disjoint_set::RankUnionPathSplitting<int16_t>>(
+  RunBenchmark<disjoint_set::RankUnionPathSplitting<int64_t>>(
       rank_union_PS_statistics_);
 }
 
 TEST_F(DisjointSetBenchmark, RankUnionPH) {
-  RunBenchmark<disjoint_set::RankUnionPathHalving<int16_t>>(
+  RunBenchmark<disjoint_set::RankUnionPathHalving<int64_t>>(
       rank_union_PH_statistics_);
 }
